+++ src/LamaPon/Components/UIInputFieldComponent.h
@@ -149,9 +149,7 @@ namespace LamaPon
             GraphicsDevice& graphics) override;
         void OnUpdate(float deltaTime) override;
         void OnRender2D(
-            DirectX::SpriteBatch& spriteBatch,
-            ID3D11ShaderResourceView*
-                whiteTexture) override;
+            const SpriteDrawContext& sprites) override;
 
     private:
         void RefreshText();