+++ src/LamaPon/Components/UISliderComponent.h
@@ -119,9 +119,7 @@ namespace LamaPon
             GraphicsDevice& graphics) override;
         void OnUpdate(float deltaTime) override;
         void OnRender2D(
-            DirectX::SpriteBatch& spriteBatch,
-            ID3D11ShaderResourceView*
-                whiteTexture) override;
+            const SpriteDrawContext& sprites) override;
 
     private:
         [[nodiscard]] float ApplyConstraints(